code/requirements.txt
coverage==7.8.2
Django==5.2.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
pytest-django==4.10.0
pytest==8.3.5
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
dotenv==0.9.9